changed: put path videodatabase methods in separate file
